+++ contrib/platform/src/com/sun/jna/platform/win32/ShlObj.java
@@ -13,7 +13,6 @@
 package com.sun.jna.platform.win32;
 
 import com.sun.jna.platform.win32.WinDef.DWORD;
-import com.sun.jna.win32.StdCallLibrary;
 
 /**
  * Ported from ShlObj.h.