+++ contrib/platform/src/com/sun/jna/platform/win32/Sspi.java
@@ -241,12 +241,12 @@ public interface Sspi extends StdCallLibrary {
         /**
          * Specifies the size, in bytes, of the buffer pointed to by the pvBuffer member.
          */
-        public NativeLong cbBuffer;
+        public int cbBuffer;
         /**
          * Bit flags that indicate the type of buffer. Must be one of the values of 
          * the SecBufferType enumeration.
          */
-        public NativeLong BufferType;
+        public int BufferType;
         /**
          * A pointer to a buffer.
          */
@@ -256,9 +256,9 @@ public interface Sspi extends StdCallLibrary {
          * Create a new SECBUFFER_EMPTY buffer.
          */
         public SecBuffer() {
-            cbBuffer = new NativeLong(0);
+            cbBuffer = 0;
             pvBuffer = null;
-            BufferType = new NativeLong(SECBUFFER_EMPTY);
+            BufferType = SECBUFFER_EMPTY;
         }
 	    
         /**
@@ -269,9 +269,9 @@ public interface Sspi extends StdCallLibrary {
          *  Buffer size, eg. MAX_TOKEN_SIZE.
          */
         public SecBuffer(int type, int size) {
-            cbBuffer = new NativeLong(size);	    	
+            cbBuffer = size;	    	
             pvBuffer = new Memory(size);
-            BufferType = new NativeLong(type);
+            BufferType = type;
             allocateMemory();
         }
 	    
@@ -283,10 +283,10 @@ public interface Sspi extends StdCallLibrary {
          *  Existing token.
          */
         public SecBuffer(int type, byte[] token) {
-            cbBuffer = new NativeLong(token.length);	    	
+            cbBuffer = token.length;	    	
             pvBuffer = new Memory(token.length);
             pvBuffer.write(0, token, 0, token.length);
-            BufferType = new NativeLong(type);
+            BufferType = type;
             allocateMemory();
         }
 	    
@@ -296,7 +296,7 @@ public interface Sspi extends StdCallLibrary {
          *  Raw buffer bytes.
          */
         public byte[] getBytes() {
-            return pvBuffer.getByteArray(0, cbBuffer.intValue());
+            return pvBuffer.getByteArray(0, cbBuffer);
         }
     }
 
@@ -305,11 +305,11 @@ public interface Sspi extends StdCallLibrary {
         /**
          * Version number.
          */
-        public NativeLong ulVersion;
+        public int ulVersion;
         /**
          * Number of buffers.
          */
-        public NativeLong cBuffers;
+        public int cBuffers;
         /**
          * Pointer to array of buffers.
          */
@@ -319,8 +319,8 @@ public interface Sspi extends StdCallLibrary {
          * Create a new SecBufferDesc with one SECBUFFER_EMPTY buffer.
          */
         public SecBufferDesc() {
-            ulVersion = new NativeLong(SECBUFFER_VERSION);
-            cBuffers = new NativeLong(1);
+            ulVersion = SECBUFFER_VERSION;
+            cBuffers = 1;
             SecBuffer.ByReference secBuffer = new SecBuffer.ByReference();
             pBuffers = (SecBuffer.ByReference[]) secBuffer.toArray(1);
             allocateMemory();
@@ -334,8 +334,8 @@ public interface Sspi extends StdCallLibrary {
          *  Initial token data.
          */
         public SecBufferDesc(int type, byte[] token) {
-            ulVersion = new NativeLong(SECBUFFER_VERSION);
-            cBuffers = new NativeLong(1);
+            ulVersion = SECBUFFER_VERSION;
+            cBuffers = 1;
             SecBuffer.ByReference secBuffer = new SecBuffer.ByReference(type, token);
             pBuffers = (SecBuffer.ByReference[]) secBuffer.toArray(1);
             allocateMemory();	    	
@@ -347,18 +347,18 @@ public interface Sspi extends StdCallLibrary {
          * @param tokenSize
          */
         public SecBufferDesc(int type, int tokenSize) {
-            ulVersion = new NativeLong(SECBUFFER_VERSION);
-            cBuffers = new NativeLong(1);
+            ulVersion = SECBUFFER_VERSION;
+            cBuffers = 1;
             SecBuffer.ByReference secBuffer = new SecBuffer.ByReference(type, tokenSize);
             pBuffers = (SecBuffer.ByReference[]) secBuffer.toArray(1);
             allocateMemory();
         }	    	
 	    
         public byte[] getBytes() {
-            if (pBuffers == null || cBuffers == null) {
+            if (pBuffers == null) {
                 throw new RuntimeException("pBuffers | cBuffers");
             }
-            if (cBuffers.intValue() == 1) {
+            if (cBuffers == 1) {
                 return pBuffers[0].getBytes();
             }	    	
             throw new RuntimeException("cBuffers > 1");