+++ contrib/platform/test/com/sun/jna/platform/win32/Shell32Test.java
@@ -15,8 +15,6 @@ package com.sun.jna.platform.win32;
 import junit.framework.TestCase;
 
 import com.sun.jna.Native;
-import com.sun.jna.Pointer;
-import com.sun.jna.Structure;
 import com.sun.jna.ptr.PointerByReference;
 
 /**
@@ -26,12 +24,6 @@ public class Shell32Test extends TestCase {
 
     public static void main(String[] args) {
         junit.textui.TestRunner.run(Shell32Test.class);
-    }
-
-    public void testStructurePacking() {
-        Structure s = new ShellAPI.SHFILEOPSTRUCT();
-        final int SIZE = Pointer.SIZE * 5 + 10; // 5 pointers, 2 ints, 1 short
-        assertEquals("Wrong structure size", SIZE, s.size());
     }
     
     public void testSHGetFolderPath() {