+++ contrib/platform/test/com/sun/jna/platform/win32/VersionTest.java
@@ -25,9 +25,9 @@ public class VersionTest extends TestCase {
 
     public void testGetFileVersion() {
         String systemRoot = System.getenv("SystemRoot");
-        File notepad = new File(systemRoot + "\\notepad.exe");
+        File explorer = new File(systemRoot + "\\explorer.exe");
 
-        int size = Version.INSTANCE.GetFileVersionInfoSize(notepad.getAbsolutePath(), null);
+        int size = Version.INSTANCE.GetFileVersionInfoSize(explorer.getAbsolutePath(), null);
         assertTrue(size > 0);
 
         Pointer buffer = Kernel32.INSTANCE.LocalAlloc(WinBase.LMEM_ZEROINIT, size);
@@ -35,7 +35,7 @@ public class VersionTest extends TestCase {
 
         try
         {
-            assertTrue(Version.INSTANCE.GetFileVersionInfo(notepad.getAbsolutePath(), 0, size, buffer));
+            assertTrue(Version.INSTANCE.GetFileVersionInfo(explorer.getAbsolutePath(), 0, size, buffer));
 
             IntByReference outputSize = new IntByReference();
             PointerByReference pointer = new PointerByReference();